Delete firebase dependencies to simplify framework roll. (#1023)
We're having trouble updating the flutter gallery to head for flutter
integration testing. Since we're not updating the gallery any further, I
believe we should simplify the problem by removing the firebase
depenencies, which are complicating the native build.

This would ultimately be required to merge this gallery into
flutter/flutter anyway.